Se modifica el tipo de variables de 32 bits

+++ test/test_eeprom.c
@@ -177,8 +177,8 @@ void test_eeprom_address_less_than_initial_address( void )
     TEST_ASSERT_EQUAL_UINT8( false, eeprom_write16( address, &write_data16 ) );
 
     /* se declaran los datos a ser escritos y leídos de 32 bits */
-    uint16_t read_data32 = 0x0;    
-    uint16_t write_data32 = 0xFFFFFFFF;    
+    uint32_t read_data32 = 0x0;    
+    uint32_t write_data32 = 0xFFFFFFFF;    
 
     /* se verifica si el valor obtenido por las funciones de lectura y escritura es false */
     TEST_ASSERT_EQUAL_UINT8( false, eeprom_read32( address, &read_data32 ) );
@@ -214,8 +214,8 @@ void test_eeprom_address_greater_than_eeprom_size( void )
     address = EEPROM_SIZE - sizeof( uint32_t ) + 1;
 
     /* se declaran los datos a ser escritos y leídos de 32 bits */
-    uint16_t read_data32 = 0x0;    
-    uint16_t write_data32 = 0xFFFFFFFF;    
+    uint32_t read_data32 = 0x0;    
+    uint32_t write_data32 = 0xFFFFFFFF;    
 
     /* se verifica si el valor obtenido por las funciones de lectura y escritura es false */
     TEST_ASSERT_EQUAL_UINT8( false, eeprom_read32( address, &read_data32 ) );
